Move class and do something weird to the comment.
