+++ config/drupal-9/drupal-9.2-deprecations.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-use DrupalRector\Rector\Deprecation\Base\MethodToMethodWithCheckRector;
+use DrupalRector\Rector\Deprecation\MethodToMethodWithCheckRector;
 use DrupalRector\Rector\ValueObject\MethodToMethodWithCheckConfiguration;
 use Rector\PHPUnit\Set\PHPUnitSetList;
